# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user_enums.UserRole:
    enum:
    - ADMIN
    - MANAGER
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleManager
  user_enums.UserStatus:
    enum:
    - ACTIVE
    - BLOCKED
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusBlocked
  users.ChangeMyPasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
  users.ChangeUserPasswordRequest:
    properties:
      newPassword:
        minLength: 8
        type: string
    required:
    - newPassword
    type: object
  users.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/user_enums.UserRole'
    required:
    - email
    - password
    - role
    type: object
  users.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.SignInResponse:
    properties:
      token:
        type: string
      userId:
        type: string
    type: object
  users.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  users.UpdateUserStatusRequest:
    properties:
      status:
        $ref: '#/definitions/user_enums.UserStatus'
    required:
    - status
    type: object
  users.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/user_enums.UserRole'
      status:
        $ref: '#/definitions/user_enums.UserStatus'
    type: object
  users_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/user_enums.UserRole'
      status:
        $ref: '#/definitions/user_enums.UserStatus'
    type: object
...
host: localhost:4005
info:
  contact: {}
  description: API for Postgresus
  termsOfService: http://swagger.io/terms/
  title: Postgresus Backend API
  version: "1.0"
paths:
...
  /users/admin/{id}/password:
    put:
      consumes:
      - application/json
      description: Change the password of a user. Only accessible by admin users.
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.ChangeUserPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change user password (Admin only)
      tags:
      - users
  /users/admin/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a user (ACTIVE/BLOCKED). Only accessible by
        admin users.
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user status (Admin only)
      tags:
      - users
  /users/admin/create-user:
    post:
      consumes:
      - application/json
      description: Create a new user with specified role. Only accessible by admin
        users.
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user (Admin only)
      tags:
      - users
  /users/admin/list:
    get:
      description: Get a list of all users in the system. Only accessible by admin
        users.
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users (Admin only)
      tags:
      - users
...
  /users/me:
    get:
      description: Get the current authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users_models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Change the authenticated user's own password
      parameters:
      - description: JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.ChangeMyPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change current user's password
      tags:
      - users
  /users/signin:
    post:
      consumes:
      - application/json
      description: Authenticate a user with email and password
      parameters:
      - description: User signin data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.SignInResponse'
        "400":
          description: Bad Request
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate a user
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User signup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Register a new user
      tags:
      - users
schemes:
- http
swagger: "2.0"